Keep the Python 3.10 legs collectible around service-only imports
Seat: lead-developer
Assignment: PH-3-UNIT-1B
